indexer: use different queues for indexing
* Adds possibility to create additional queues for bulk indexing.
* Groups all custom command lines are now under `invenio reroils`.
