+++ .coveragerc
@@ -3,7 +3,11 @@ relative_files = True
 omit =
     rero_ils/dojson/cli.py
     rero_ils/modules/apiharvester/cli.py
-    rero_ils/modules/cli.py
+    rero_ils/modules/cli/fixtures.py
+    rero_ils/modules/cli/index.py
+    rero_ils/modules/cli/reroils.py
+    rero_ils/modules/cli/users.py
+    rero_ils/modules/cli/utils.py
     rero_ils/modules/collections/cli.py
     rero_ils/modules/ebooks/cli.py
     rero_ils/modules/holdings/cli.py
