+++ rero_ils/modules/cli/users.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 #
 # RERO ILS
-# Copyright (C) 2019 RERO
+# Copyright (C) 2021 RERO
+# Copyright (C) 2020 UCLouvain
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
@@ -17,43 +18,33 @@
 
 """Click command-line utilities."""
 
-import json
+from __future__ import absolute_import, print_function
 
 import click
 from flask import current_app
+from flask.cli import with_appcontext
+from flask_security.confirmable import confirm_user
+from invenio_accounts.cli import commit, users
 from werkzeug.local import LocalProxy
 
 _datastore = LocalProxy(lambda: current_app.extensions['security'].datastore)
 
 
-@click.command('reverse')
-def reverse():
-    """Reverse the order of the data."""
-    def processor(iterator):
-        items = []
-        for item in iterator:
-            items.append(item)
-        items.reverse()
-        return items
-
-    return processor
+@click.group()
+def users():
+    """Index management commands."""
 
 
-@click.command('head')
-@click.argument('max', type=click.INT,)
-def head(max):
-    """Take only the first max items."""
-    def processor(iterator):
-        n = 0
-        for item in iterator:
-            if n >= max:
-                raise StopIteration
-            n += 1
-            yield item
-
-    return processor
-
-
-def pretty_json_dump(iterator):
-    """Dump JSON from iteraror."""
-    return json.dumps(list(iterator), indent=2)
+@users.command('confirm')
+@click.argument('user')
+@with_appcontext
+@commit
+def manual_confirm_user(user):
+    """Confirm a user."""
+    user_obj = _datastore.get_user(user)
+    if user_obj is None:
+        raise click.UsageError('ERROR: User not found.')
+    if confirm_user(user_obj):
+        click.secho('User "%s" has been confirmed.' % user, fg='green')
+    else:
+        click.secho('User "%s" was already confirmed.' % user, fg='yellow')
