+++ rero_ils/modules/cli/utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # RERO ILS
-# Copyright (C) 2019 RERO
+# Copyright (C) 2021 RERO
 # Copyright (C) 2020 UCLouvain
 #
 # This program is free software: you can redistribute it and/or modify
@@ -30,33 +30,24 @@ import re
 import sys
 import traceback
 from collections import OrderedDict
-from datetime import datetime
 from glob import glob
 from pprint import pprint
 from time import sleep
-from uuid import uuid4
 
 import click
-import dateparser
 import yaml
 from celery import current_app as current_celery
 from dojson.contrib.marc21.utils import create_record
 from elasticsearch_dsl.query import Q
 from flask import current_app
 from flask.cli import with_appcontext
-from flask_security.confirmable import confirm_user
-from invenio_accounts.cli import commit, users
 from invenio_db import db
 from invenio_jsonschemas.proxies import current_jsonschemas
 from invenio_oauth2server.cli import process_scopes, process_user
 from invenio_oauth2server.models import Client, Token
 from invenio_pidstore.models import PersistentIdentifier, PIDStatus
 from invenio_records.api import Record
-from invenio_records.models import RecordMetadata
-from invenio_records_rest.utils import obj_or_import_string
-from invenio_search.cli import es_version_check
-from invenio_search.proxies import current_search, current_search_client
-from jsonpatch import make_patch
+from invenio_search.proxies import current_search
 from jsonschema import validate
 from jsonschema.exceptions import ValidationError
 from lxml import etree
@@ -65,85 +56,20 @@ from werkzeug.security import gen_salt
 
 from rero_ils.modules.locations.api import Location
 
-from .api import IlsRecordsIndexer
-from .collections.cli import create_collections
-from .contributions.api import Contribution
-from .contributions.tasks import create_mef_record_online
-from .documents.api import Document, DocumentsIndexer, DocumentsSearch
-from .documents.dojson.contrib.marc21tojson import marc21
-from .documents.views import get_cover_art
-from .holdings.cli import create_patterns
-from .ill_requests.cli import create_ill_requests
-from .items.api import Item
-from .items.cli import create_items, reindex_items
-from .libraries.api import Library
-from .loans.cli import create_loans, load_virtua_transactions
-from .monitoring import Monitoring
-from .operation_logs.cli import create_operation_logs, \
-    destroy_operation_logs, dump_operation_logs
-from .patrons.cli import import_users, users_validate
-from .selfcheck.cli import create_terminal, list_terminal, update_terminal
-from .tasks import process_bulk_queue
-from .utils import JsonWriter, bulk_load_metadata, bulk_load_pids, \
-    bulk_load_pidstore, bulk_save_metadata, bulk_save_pids, \
-    bulk_save_pidstore, csv_metadata_line, csv_pidstore_line, \
-    extracted_data_from_ref, get_record_class_from_schema_or_pid_type, \
-    number_records_in_file, read_json_record, read_xml_record
-from ..modules.providers import append_fixtures_new_identifiers
-from ..modules.utils import get_schema_for_resource
+from ..documents.api import Document, DocumentsIndexer, DocumentsSearch
+from ..documents.dojson.contrib.marc21tojson import marc21
+from ..documents.views import get_cover_art
+from ..items.api import Item
+from ..libraries.api import Library
+from ..patrons.cli import users_validate
+from ..selfcheck.cli import create_terminal, list_terminal, update_terminal
+from ..tasks import process_bulk_queue
+from ..utils import JsonWriter, extracted_data_from_ref, \
+    get_record_class_from_schema_or_pid_type, get_schema_for_resource, \
+    read_json_record, read_xml_record
 
 _datastore = LocalProxy(lambda: current_app.extensions['security'].datastore)
 _records_state = LocalProxy(lambda: current_app.extensions['invenio-records'])
-
-
-def abort_if_false(ctx, param, value):
-    """Abort command is value is False."""
-    if not value:
-        ctx.abort()
-
-
-@click.group()
-def fixtures():
-    """Fixtures management commands."""
-
-
-fixtures.add_command(import_users)
-fixtures.add_command(create_items)
-fixtures.add_command(reindex_items)
-fixtures.add_command(create_loans)
-fixtures.add_command(load_virtua_transactions)
-fixtures.add_command(create_patterns)
-fixtures.add_command(create_ill_requests)
-fixtures.add_command(create_collections)
-fixtures.add_command(create_operation_logs)
-fixtures.add_command(dump_operation_logs)
-fixtures.add_command(destroy_operation_logs)
-
-
-@users.command('confirm')
-@click.argument('user')
-@with_appcontext
-@commit
-def manual_confirm_user(user):
-    """Confirm a user."""
-    user_obj = _datastore.get_user(user)
-    if user_obj is None:
-        raise click.UsageError('ERROR: User not found.')
-    if confirm_user(user_obj):
-        click.secho('User "%s" has been confirmed.' % user, fg='green')
-    else:
-        click.secho('User "%s" was already confirmed.' % user, fg='yellow')
-
-
-@click.group()
-def utils():
-    """Misc management commands."""
-
-
-utils.add_command(users_validate)
-utils.add_command(create_terminal)
-utils.add_command(list_terminal)
-utils.add_command(update_terminal)
 
 
 def queue_count():
@@ -182,6 +108,17 @@ def wait_empty_tasks(delay, verbose=False):
         count = queue_count()
         sleep(5)
         count += queue_count()
+
+
+@click.group()
+def utils():
+    """Utils management commands."""
+
+
+utils.add_command(users_validate)
+utils.add_command(create_terminal)
+utils.add_command(list_terminal)
+utils.add_command(update_terminal)
 
 
 @utils.command('wait_empty_tasks')
@@ -280,242 +217,6 @@ def schedules():
     for key, value in celery_ext.celery.conf.beat_schedule.items():
         click.echo(key + '\t', nl=False)
         click.echo(value)
-
-
-@utils.command('init_index')
-@click.option('--force', is_flag=True, default=False)
-@with_appcontext
-@es_version_check
-def init_index(force):
-    """Initialize registered templates, aliases and mappings."""
-    # TODO: to remove once it is fixed in invenio-search module
-    click.secho('Putting templates...', fg='green', bold=True, file=sys.stderr)
-    with click.progressbar(
-            current_search.put_templates(ignore=[400] if force else None),
-            length=len(current_search.templates)) as bar:
-        for response in bar:
-            bar.label = response
-    click.secho('Creating indexes...', fg='green', bold=True, file=sys.stderr)
-    with click.progressbar(
-            current_search.create(ignore=[400] if force else None),
-            length=len(current_search.mappings)) as bar:
-        for name, response in bar:
-            bar.label = name
-
-
-@utils.command('switch_index')
-@with_appcontext
-@es_version_check
-@click.argument('old')
-@click.argument('new')
-def switch_index(old, new):
-    """Switch index using the elasticsearch aliases.
-
-    :param old: full name of the old index
-    :param new: full name of the fresh created index
-    """
-    aliases = current_search_client.indices.get_alias().get(old)\
-        .get('aliases').keys()
-    for alias in aliases:
-        current_search_client.indices.put_alias(new, alias)
-        current_search_client.indices.delete_alias(old, alias)
-    click.secho('Sucessfully switched.', fg='green')
-
-
-@utils.command('create_index')
-@with_appcontext
-@es_version_check
-@click.option(
-    '-t', '--templates/--no-templates', 'templates', is_flag=True,
-    default=True)
-@click.option(
-    '-v', '--verbose/--no-verbose', 'verbose', is_flag=True, default=False)
-@click.argument('resource')
-@click.argument('index')
-def create_index(resource, index, verbose, templates):
-    """Create a new index based on the mapping of a given resource.
-
-    :param resource: the resource such as documents.
-    :param index: the index name such as documents-document-v0.0.1-20211014
-    :param verbose: display addtional message.
-    :param templates: update also the es templates.
-    """
-    if templates:
-        tbody = current_search_client.indices.get_template()
-        for tmpl in current_search.put_templates():
-            click.secho(f'file:{tmpl[0]}, ok: {tmpl[1]}', fg='green')
-            new_tbody = current_search_client.indices.get_template()
-            patch = make_patch(new_tbody, tbody)
-            if patch:
-                click.secho('Templates are updated.', fg='green')
-                if verbose:
-                    click.secho('Diff in templates', fg='green')
-                    click.echo(patch)
-            else:
-                click.secho('Templates did not changed.', fg='yellow')
-
-    f_mapping = [
-        v for v in current_search.aliases.get(resource).values()].pop()
-    mapping = json.load(open(f_mapping))
-    current_search_client.indices.create(index, mapping)
-    click.secho(f'Index {index} has been created.', fg='green')
-
-
-@utils.command('update_mapping')
-@click.option('--aliases', '-a', multiple=True, help='all if not specified')
-@with_appcontext
-@es_version_check
-def update_mapping(aliases):
-    """Update the mapping of a given alias."""
-    if not aliases:
-        aliases = current_search.aliases.keys()
-    for alias in aliases:
-        for index, f_mapping in iter(
-            current_search.aliases.get(alias).items()
-        ):
-            mapping = json.load(open(f_mapping))
-            try:
-                res = current_search_client.indices.put_mapping(
-                    mapping.get('mappings'), index)
-            except Exception as excep:
-                click.secho(
-                    f'error: {excep}', fg='red')
-            if res.get('acknowledged'):
-                click.secho(
-                    f'index: {index} has been sucessfully updated', fg='green')
-            else:
-                click.secho(
-                    f'error: {res}', fg='red')
-
-
-@fixtures.command('create')
-@click.option('-u', '--create_or_update', 'create_or_update', is_flag=True,
-              default=False)
-@click.option('-a', '--append', 'append', is_flag=True, default=False)
-@click.option('-r', '--reindex', 'reindex', is_flag=True, default=False)
-@click.option('-c', '--dbcommit', 'dbcommit', is_flag=True, default=False)
-@click.option('-C', '--commit', 'commit', default=100000)
-@click.option('-v', '--verbose', 'verbose', is_flag=True, default=True)
-@click.option('-d', '--debug', 'debug', is_flag=True, default=False)
-@click.option('-s', '--schema', 'schema', default=None)
-@click.option('-t', '--pid_type', 'pid_type', default=None)
-@click.option('-l', '--lazy', 'lazy', is_flag=True, default=False)
-@click.option('-o', '--dont-stop', 'dont_stop_on_error',
-              is_flag=True, default=False)
-@click.option('-P', '--pid-check', 'pid_check',
-              is_flag=True, default=False)
-@click.option('-e', '--save_errors', 'save_errors', type=click.File('w'))
-@click.argument('infile', type=click.File('r'), default=sys.stdin)
-@with_appcontext
-def create(infile, create_or_update, append, reindex, dbcommit, commit,
-           verbose, debug, schema, pid_type, lazy, dont_stop_on_error,
-           pid_check, save_errors):
-    """Load REROILS record.
-
-    :param infile: Json file
-    :param create_or_update: to update or create records.
-    :param append: appends pids to database
-    :param reindex: reindex record by record
-    :param dbcommit: commit record to database
-    :param commit: commit to database every count records
-    :param pid_type: record type
-    :param schema: recoord schema
-    :param lazy: lazy reads file
-    :param dont_stop_on_error: don't stop on error
-    :param pidcheck: check pids
-    :param save_errors: save error records to file
-    """
-    click.secho(
-        f'Loading {pid_type} records from {infile.name}.',
-        fg='green'
-    )
-
-    record_class = get_record_class_from_schema_or_pid_type(pid_type=pid_type)
-
-    if save_errors:
-        errors = 0
-        name, ext = os.path.splitext(infile.name)
-        err_file_name = f'{name}_errors{ext}'
-        error_file = JsonWriter(err_file_name)
-
-    pids = []
-    if lazy:
-        # try to lazy read json file (slower, better memory management)
-        records = read_json_record(infile)
-    else:
-        # load everything in memory (faster, bad memory management)
-        records = json.load(infile)
-    count = 0
-    for count, record in enumerate(records, 1):
-        if schema:
-            record['$schema'] = schema
-        try:
-            pid = record.get('pid')
-            msg = 'created'
-            db_record = record_class.get_record_by_pid(pid)
-            if create_or_update and db_record:
-                # case when record already exist in database
-                db_record = record_class.get_record_by_pid(pid)
-                rec = db_record.update(
-                        record, dbcommit=dbcommit, reindex=reindex)
-                msg = 'updated'
-            elif create_or_update and pid and not db_record \
-                    and record_class.record_pid_exists(pid):
-                # case when record not in db but pid is reserved
-                presist_id = PersistentIdentifier.get(
-                    record_class.provider.pid_type, pid)
-                rec = record_class.create(
-                    record, dbcommit=dbcommit, reindex=reindex)
-                if presist_id.status != PIDStatus.REGISTERED:
-                    presist_id.register()
-                    presist_id.assign(record_class.object_type, rec.id)
-                msg = 'created'
-            else:
-                # case when record and pid are not in db
-                rec = record_class.create(
-                        record, dbcommit=dbcommit, reindex=reindex,
-                        pidcheck=pid_check)
-                if append:
-                    pids.append(rec.pid)
-            if verbose:
-                click.echo(
-                    f'{count: <8} {pid_type} {msg} {rec.pid}:{rec.id}')
-
-        except Exception as err:
-            pid = record.get('pid', '???')
-            click.secho(
-                f'{count: <8} {type} create error {pid}: {err}',
-                fg='red'
-            )
-            if debug:
-                traceback.print_exc()
-
-            if save_errors:
-                error_file.write(record)
-            if not dont_stop_on_error:
-                sys.exit(1)
-        db.session.flush()
-        if count > 0 and count % commit == 0:
-            if verbose:
-                click.echo(f'DB commit: {count}')
-            db.session.commit()
-    click.echo(f'DB commit: {count}')
-    db.session.commit()
-
-    if append:
-        click.secho(f'Append fixtures new identifiers: {len(pids)}')
-        identifier = record_class.provider.identifier
-        try:
-            append_fixtures_new_identifiers(
-                identifier,
-                sorted(pids, key=lambda x: int(x)),
-                pid_type
-            )
-        except Exception as err:
-            click.secho(
-                f'ERROR append fixtures new identifiers: {err}',
-                fg='red'
-            )
 
 
 @utils.command('validate_documents_with_items')
@@ -727,96 +428,6 @@ def create_documents_with_items(infile, lazy, dont_stop_on_error, save_errors,
                 f'         - {item["count"]: <8}'
                 f' doc:{item["rero_ils_pid"]}'
                 f' item created {item_rec.pid}:{item_rec.id}')
-
-
-@fixtures.command('count')
-@click.option('-l', '--lazy', 'lazy', is_flag=True, default=False)
-@click.argument('infile', type=click.File('r'), default=sys.stdin)
-def count_cli(infile, lazy):
-    """Count records in file.
-
-    :param infile: Json file
-    :param lazy: lazy reads file
-    :return: count of records
-    """
-    click.secho(
-        f'Count records from {infile.name}.',
-        fg='green'
-    )
-    if lazy:
-        # try to lazy read json file (slower, better memory management)
-        records = read_json_record(infile)
-    else:
-        # load everything in memory (faster, bad memory management)
-        records = json.load(infile)
-    count = 0
-    for record in records:
-        count += 1
-    click.echo(f'Count: {count}')
-
-
-@fixtures.command('get_all_mef_records')
-@click.argument('infile', type=click.File('r'), default=sys.stdin)
-@click.option('-l', '--lazy', 'lazy', is_flag=True, default=False,
-              help="lazy reads file")
-@click.option('-k', '--enqueue', 'enqueue', is_flag=True, default=False,
-              help="Enqueue record creation.")
-@click.option('-v', '--verbose', 'verbose', is_flag=True, default=True,
-              help='verbose')
-@click.option('-w', '--wait', 'wait', is_flag=True, default=False,
-              help="wait for enqueued tasks to finish")
-@click.option('-o', '--out_file', 'outfile_name', default=None)
-@with_appcontext
-def get_all_mef_records(infile, lazy, verbose, enqueue, wait, outfile_name):
-    """Get all contributions for given document file."""
-    click.secho(
-        f'Get all contributions for {infile.name}.',
-        fg='green'
-    )
-    if outfile_name:
-        outfile = JsonWriter(outfile_name)
-        contribution_schema = get_schema_for_resource('cont')
-        click.secho(f'Write to: {outfile_name}.')
-    if lazy:
-        # try to lazy read json file (slower, better memory management)
-        records = read_json_record(infile)
-    else:
-        # load everything in memory (faster, bad memory management)
-        records = json.load(infile)
-    count = 0
-    refs = {}
-    for count, record in enumerate(records, 1):
-        for contribution in record.get('contribution', []):
-            ref = contribution['agent'].get('$ref')
-            if ref and not refs.get(ref):
-                refs[ref] = 1
-                if outfile_name:
-                    try:
-                        ref_split = ref.split('/')
-                        ref_type = ref_split[-2]
-                        ref_pid = ref_split[-1]
-                        data = Contribution._get_mef_data_by_type(
-                            pid=ref_pid,
-                            pid_type=ref_type
-                        )
-                        metadata = data['metadata']
-                        metadata['$schema'] = contribution_schema
-                        outfile.write(metadata)
-                        msg = 'ok'
-                    except Exception as err:
-                        msg = err
-                else:
-                    if enqueue:
-                        msg = create_mef_record_online.delay(ref)
-                    else:
-                        pid, online = create_mef_record_online(ref)
-                        msg = f'contribution pid: {pid} {online}'
-                if verbose:
-                    click.echo(f"{count:<10}ref: {ref}\t{msg}")
-    else:
-        if enqueue and wait:
-            wait_empty_tasks(delay=3, verbose=True)
-    click.echo(f'Count refs: {count}')
 
 
 @utils.command('check_license')
@@ -1424,175 +1035,6 @@ def reserve_pid_range(pid_type, records_number, unused):
             db.session.commit()
 
 
-@utils.command('runindex')
-@click.option(
-    '--delayed', '-d', is_flag=True, help='Run indexing in background.')
-@click.option(
-    '--concurrency', '-c', default=1, type=int,
-    help='Number of concurrent indexing tasks to start.')
-@click.option(
-    '--with_stats', is_flag=True, default=False,
-    help='report number of successful and list failed error response.')
-@click.option('--queue', '-q', type=str,
-              help='Name of the celery queue used to put the tasks into.')
-@click.option('--version-type', help='Elasticsearch version type to use.')
-@click.option(
-    '--raise-on-error/--skip-errors', default=True,
-    help='Controls if Elasticsearch bulk indexing errors raise an exception.')
-@with_appcontext
-def run(delayed, concurrency, with_stats, version_type=None, queue=None,
-        raise_on_error=True):
-    """Run bulk record indexing."""
-    if delayed:
-        celery_kwargs = {
-            'kwargs': {
-                'version_type': version_type,
-                'es_bulk_kwargs': {'raise_on_error': raise_on_error},
-                'stats_only': not with_stats
-            }
-        }
-        click.secho(
-            f'Starting {concurrency} tasks for indexing records...',
-            fg='green'
-        )
-        if queue is not None:
-            celery_kwargs.update({'queue': queue})
-        for c in range(0, concurrency):
-            process_id = process_bulk_queue.apply_async(**celery_kwargs)
-            click.secho(f'index async: {process_id}', fg='yellow')
-
-    else:
-        click.secho('Indexing records...', fg='green')
-        indexed, error = IlsRecordsIndexer(version_type=version_type)\
-            .process_bulk_queue(
-                es_bulk_kwargs={'raise_on_error': raise_on_error},
-                stats_only=not with_stats
-            )
-        click.secho(f'indexed: {indexed}, error: {error}', fg='yellow')
-
-
-@utils.command('reindex')
-@click.option('--yes-i-know', is_flag=True, callback=abort_if_false,
-              expose_value=False,
-              prompt='Do you really want to reindex records?')
-@click.option('-t', '--pid-types', multiple=True)
-@click.option('-n', '--no-info', 'no_info', is_flag=True, default=True)
-@click.option('-f', '--from_date', 'from_date')
-@click.option('-u', '--until_date', 'until_date')
-@click.option('-d', '--direct', 'direct', is_flag=True, default=False)
-@click.option('-c', '--count', 'count', is_flag=True, default=False)
-@click.option('-i', '--index', 'index')
-@with_appcontext
-def reindex(pid_types, no_info, from_date, until_date, direct, count, index):
-    """Reindex records.
-
-    :param pid_type: Pid type.
-    :param no_info: Display no info.
-    :param from_date: Index records from date.
-    :param until_date: Index records until date.
-    :param direct: Use record class for indexing.
-    :param count: Do not index, display only counts.
-    :param index: Index name to index.
-    """
-    endpoints = current_app.config.get('RECORDS_REST_ENDPOINTS')
-    if not pid_types:
-        pid_types = [endpoint for endpoint in endpoints]
-    for pid_type in pid_types:
-        if pid_type in endpoints:
-            msg = f'Sending {pid_type} to indexing queue: '
-            if count:
-                msg = f'Count {pid_type:>6}: '
-            click.secho(msg, fg='green', nl=False)
-            query = None
-            record_cls = obj_or_import_string(
-                endpoints[pid_type].get('record_class'))
-            if from_date or until_date:
-                model_cls = record_cls.model_cls
-                if model_cls != RecordMetadata:
-                    query = model_cls.query \
-                        .filter(model_cls.is_deleted.is_(False)) \
-                        .with_entities(model_cls.id) \
-                        .order_by(model_cls.created)
-                    if from_date:
-                        query = query.filter(
-                            model_cls.updated > dateparser.parse(from_date))
-                    if until_date:
-                        query = query.filter(
-                            model_cls.updated <= dateparser.parse(until_date))
-            else:
-                query = PersistentIdentifier.query \
-                    .filter_by(object_type='rec', status=PIDStatus.REGISTERED)\
-                    .filter_by(pid_type=pid_type) \
-                    .with_entities(PersistentIdentifier.object_uuid)
-            if query:
-                click.echo(f'{query.count()}')
-                if not count:
-                    if direct:
-                        for idx, id in enumerate((x[0] for x in query), 1):
-                            msg = f'{idx}\t{id}\t'
-                            try:
-                                rec = record_cls.get_record_by_id(id)
-                                msg += f'{rec.pid}'
-                                rec.reindex()
-                            except Exception as err:
-                                msg += f'\t{err}'
-                            click.echo(msg)
-                    else:
-                        IlsRecordsIndexer().bulk_index(
-                            (x[0] for x in query),
-                            doc_type=pid_type, index=index)
-            else:
-                click.echo('Can not index by date.')
-        else:
-            click.secho(f'ERROR type does not exist: {pid_type}', fg='red')
-    if no_info and not direct and not count:
-        click.secho(
-            'Execute "runindex" command to process the queue!',
-            fg='yellow'
-        )
-
-
-@utils.command('reindex_missing')
-@click.option('-t', '--pid-types', multiple=True, required=True)
-@click.option('-v', '--verbose', 'verbose', is_flag=True, default=False)
-@with_appcontext
-def reindex_missing(pid_types, verbose):
-    """Index all missing records.
-
-    :param pid_type: Pid type.
-    """
-    for p_type in pid_types:
-        click.secho(
-            f'Indexing missing {p_type}: ',
-            fg='green',
-            nl=False
-        )
-        record_class = get_record_class_from_schema_or_pid_type(
-            pid_type=p_type
-        )
-        if not record_class:
-            click.secho(
-                'ERROR pid type does not exist!',
-                fg='red',
-            )
-            continue
-        pids_es, pids_db, pids_es_double, index = \
-            Monitoring.get_es_db_missing_pids(p_type)
-        click.secho(
-            f'{len(pids_db)}',
-            fg='green',
-        )
-        for idx, pid in enumerate(pids_db, 1):
-            record = record_class.get_record_by_pid(pid)
-            if record:
-                record.reindex()
-                if verbose:
-                    click.secho(f'{idx}\t{p_type}\t{pid}')
-            else:
-                if verbose:
-                    click.secho(f'NOT FOUND: {idx}\t{p_type}\t{pid}', fg='red')
-
-
 @utils.command('check_pid_dependencies')
 @click.option('-i', '--dependency_file', 'dependency_file',
               type=click.File('r'), default='./data/pid_dependencies_big.json')
@@ -1947,223 +1389,3 @@ def add_cover_urls(verbose):
         url = get_cover_art(record=record, save_cover_url=True)
         if verbose:
             click.echo(f'{idx}:\tdocument: {pid}\t{url}')
-
-
-@fixtures.command('create_csv')
-@click.argument('record_type')
-@click.argument('json_file')
-@click.argument('output_directory')
-@click.option('-l', '--lazy', 'lazy', is_flag=True, default=False)
-@click.option('-v', '--verbose', 'verbose', is_flag=True, default=False)
-@click.option('-p', '--create_pid', 'create_pid', is_flag=True, default=False)
-@with_appcontext
-def create_csv(record_type, json_file, output_directory, lazy, verbose,
-               create_pid):
-    """Create csv files from json.
-
-    :param verbose: Verbose.
-    """
-    click.secho(
-        f"Create CSV files for: {record_type} from: {json_file}",
-        fg='green'
-    )
-
-    path = current_jsonschemas.url_to_path(
-        get_schema_for_resource(record_type)
-    )
-    add_schema = get_schema_for_resource(record_type)
-    schema = current_jsonschemas.get_schema(path=path)
-    schema = _records_state.replace_refs(schema)
-    count = 0
-    errors_count = 0
-    with open(json_file) as infile:
-        if lazy:
-            # try to lazy read json file (slower, better memory management)
-            records = read_json_record(infile)
-        else:
-            # load everything in memory (faster, bad memory management)
-            records = json.load(infile)
-
-        file_name_pidstore = os.path.join(
-            output_directory, f'{record_type}_pidstore.csv')
-        click.secho(f'\t{file_name_pidstore}', fg='green')
-        file_pidstore = open(file_name_pidstore, 'w')
-        file_name_metadata = os.path.join(
-            output_directory, f'{record_type}_metadata.csv'
-        )
-        click.secho(f'\t{file_name_metadata}', fg='green')
-        file_metadata = open(file_name_metadata, 'w')
-        file_name_pids = os.path.join(
-            output_directory, f'{record_type}_pids.csv')
-        click.secho(f'\t{file_name_pids}', fg='green')
-        file_pids = open(file_name_pids, 'w')
-        file_name_errors = os.path.join(
-            output_directory, f'{record_type}_errors.json')
-        file_errors = open(file_name_errors, 'w')
-        file_errors.write('[')
-
-        for count, record in enumerate(records, 1):
-            pid = record.get('pid')
-            if create_pid:
-                pid = str(count)
-                record['pid'] = pid
-            uuid = str(uuid4())
-            if verbose:
-                click.secho(f'{count}\t{record_type}\t{pid}:{uuid}')
-            date = str(datetime.utcnow())
-            record['$schema'] = add_schema
-            try:
-                validate(record, schema)
-                file_metadata.write(
-                    csv_metadata_line(record, uuid, date)
-                )
-                file_pidstore.write(
-                    csv_pidstore_line(record_type, pid, uuid, date)
-                )
-                file_pids.write(pid + '\n')
-            except Exception as err:
-                click.secho(
-                    f'{count}\t{record_type}: Error validate in record: ',
-                    fg='red')
-                click.secho(str(err))
-                if errors_count > 0:
-                    file_errors.write(',')
-                errors_count += 1
-                file_errors.write('\n')
-                for line in json.dumps(record, indent=2).split('\n'):
-                    file_errors.write('  ' + line + '\n')
-
-        file_pidstore.close()
-        file_metadata.close()
-        file_pids.close()
-        file_errors.write('\n]')
-        file_errors.close()
-    if errors_count == 0:
-        os.remove(file_name_errors)
-    click.secho(
-        f'Created: {count-errors_count} Errors: {errors_count}',
-        fg='yellow'
-    )
-
-
-@fixtures.command('bulk_load')
-@click.argument('record_type')
-@click.argument('csv_metadata_file')
-@click.option('-c', '--bulk_count', 'bulkcount', default=0, type=int,
-              help='Set the bulk load chunk size.')
-@click.option('-r', '--reindex', 'reindex', help='add record to reindex.',
-              is_flag=True, default=False)
-@click.option('-v', '--verbose', 'verbose', is_flag=True, default=False)
-@with_appcontext
-def bulk_load(record_type, csv_metadata_file, bulkcount, reindex, verbose):
-    """Agency record management.
-
-    :param csv_metadata_file: metadata: CSV file.
-    :param bulk_count: Set the bulk load chunk size.
-    :param reindex: add record to reindex.
-    :param verbose: Verbose.
-    """
-    if bulkcount > 0:
-        bulk_count = bulkcount
-    else:
-        bulk_count = current_app.config.get('BULK_CHUNK_COUNT', 100000)
-
-    message = f'Load {record_type} CSV files into database.'
-    click.secho(message, fg='green')
-    file_name_metadata = csv_metadata_file
-    file_name_pidstore = file_name_metadata.replace('metadata', 'pidstore')
-    file_name_pids = file_name_metadata.replace('metadata', 'pids')
-
-    record_counts = number_records_in_file(file_name_pidstore, 'csv')
-    message = f'  Number of records to load: {record_counts}'
-    click.secho(message, fg='green')
-
-    click.secho(f'  Load pids: {file_name_pids}')
-    bulk_load_pids(pid_type=record_type, ids=file_name_pids,
-                   bulk_count=bulk_count, verbose=verbose)
-    click.secho(f'  Load pidstore: {file_name_pidstore}')
-    bulk_load_pidstore(pid_type=record_type, pidstore=file_name_pidstore,
-                       bulk_count=bulk_count, verbose=verbose)
-    click.secho(f'  Load metatada: {file_name_metadata}')
-    bulk_load_metadata(pid_type=record_type, metadata=file_name_metadata,
-                       bulk_count=bulk_count, verbose=verbose, reindex=reindex)
-
-
-@fixtures.command('bulk_save')
-@click.argument('output_directory')
-@click.option('-t', '--pid_types', multiple=True, default=['all'])
-@click.option('-d', '--deployment', 'deployment', is_flag=True, default=False)
-@click.option('-v', '--verbose', 'verbose', is_flag=True, default=False)
-@with_appcontext
-def bulk_save(pid_types, output_directory, deployment, verbose):
-    """Record dump.
-
-    :param pid_type: Records to export.
-        default=//all//
-    :param verbose: Verbose.
-    """
-    file_name_tmp_pidstore = os.path.join(
-        output_directory,
-        'tmp_pidstore.csv'
-    )
-    try:
-        os.remove(file_name_tmp_pidstore)
-    except OSError:
-        pass
-
-    all_pid_types = []
-    endpoints = current_app.config.get('RECORDS_REST_ENDPOINTS')
-    for endpoint in endpoints:
-        all_pid_types.append(endpoint)
-    if pid_types[0] == 'all':
-        pid_types = all_pid_types
-
-    for p_type in pid_types:
-        if p_type not in all_pid_types:
-            click.secho(
-                f'Error {p_type} does not exist!',
-                fg='red'
-            )
-            continue
-        # TODO: do we have to save loanid and how we can save it?
-        if p_type == 'loanid':
-            continue
-        click.secho(
-            f'Save {p_type} CSV files to directory: {output_directory}',
-            fg='green'
-        )
-        file_prefix = endpoints[p_type].get('search_index')
-        if p_type in ['doc', 'hold', 'item', 'count']:
-            if deployment:
-                file_prefix += '_big'
-            else:
-                file_prefix += '_small'
-        file_name_metadata = os.path.join(
-            output_directory,
-            f'{file_prefix}_metadata.csv'
-        )
-        bulk_save_metadata(pid_type=p_type, file_name=file_name_metadata,
-                           verbose=verbose)
-        file_name_pidstore = os.path.join(
-            output_directory,
-            f'{file_prefix}_pidstore.csv'
-        )
-        count = bulk_save_pidstore(pid_type=p_type,
-                                   file_name=file_name_pidstore,
-                                   file_name_tmp=file_name_tmp_pidstore,
-                                   verbose=verbose)
-
-        file_name_pids = os.path.join(
-            output_directory,
-            f'{file_prefix}_pids.csv'
-        )
-        bulk_save_pids(pid_type=p_type, file_name=file_name_pids,
-                       verbose=verbose)
-        click.secho(
-            f'Saved records: {count}',
-            fg='yellow'
-        )
-    try:
-        os.remove(file_name_tmp_pidstore)
-    except OSError:
-        pass
