+++ scripts/update
@@ -69,5 +69,5 @@ script_path=$(dirname "$0")
 
 invenio alembic upgrade
 # TODO: this failed because we have two loan indexes
-# invenio utils init_index --force
-invenio utils update_mapping
+# invenio reroils utils init_index --force
+invenio reroils utils update_mapping
