+++ setup.py
@@ -141,14 +141,13 @@ setup(
             '.contrib.unimarctojson:unimarc',
         ],
         'flask.commands': [
+            'reroils = rero_ils.modules.cli.reroils:reroils',
             'apiharvester = rero_ils.modules.apiharvester.cli:apiharvester',
-            'fixtures = rero_ils.modules.cli:fixtures',
             'monitoring = rero_ils.modules.monitoring:monitoring',
             'notifications = rero_ils.modules.notifications.cli:notifications',
             'stats = rero_ils.modules.stats.cli:stats',
             'oaiharvester = rero_ils.modules.ebooks.cli:oaiharvester',
             'scheduler = rero_ils.schedulers:scheduler',
-            'utils = rero_ils.modules.cli:utils',
         ],
         'invenio_db.alembic': [
             'rero_ils = rero_ils:alembic',
