+++ tests/api/patrons/test_patrons_views.py
@@ -22,7 +22,7 @@ from copy import deepcopy
 from invenio_accounts.testutils import login_user_via_session
 from utils import postdata
 
-from rero_ils.modules.cli import create_personal
+from rero_ils.modules.cli.utils import create_personal
 from rero_ils.modules.items.models import ItemStatus
 from rero_ils.modules.loans.api import LoanAction
